fog-donors: Added the fog data from a Paypal log
Hopefully, we can adjust the relevant PHP bits to load from the filesystem
and not via HTTP.

Essentially I used the following to grep the data out:

   < /tmp/GNOME-Paypal-2012.csv grep 'friends@gnome.org' | grep -v 'Payment Sent' | grep -v '"List name in the donors page","No"' | grep "annually" | csvfilter -f 3 | tr -d '"' | sort -u | iconv --from-code windows-1252   > /tmp/gnome-donors-annually-2012.txt

Or

   < /tmp/GNOME-Paypal-2012.csv grep 'friends@gnome.org' | grep -v 'Payment Sent' | grep -v '"List name in the donors page","No"' | grep "Associate" | csvfilter -f 3 | tr -d '"' | sort -u | iconv --from-code windows-1252   >  /tmp/gnome-donors-2012-associate.txt

It's a bit flawed since we have some multiline donations, i.e. with comments.
Also, the data is not verified, i.e. people may have donated on a sponsor
level without actually donating the amount of money that was necessary.
At a later stage, we might provide more accurate data.

Also, I can not reproduce some data currently on the foundation's web page.
So few people are now gone from the listing. Once I understand who is
eligible to belisted, it's of course no problem to be put in the text files.