Added a special Friends of GNOME template
The purpose is to unclutter the content from within wordpress
and be able to actually program the FoG as we have many repetitions
right now.

If a page is called "friendsnew" wordpress will use this template
to render the page.

As we have all content in this template right now, the actual page
can be empty. But this is only the very first version.

The template came from index.php. It may have been better to "subclass"
the index.php, but I don't know how to do that.

In the future, the repetitive stuff should go in the template and the
actual content, like the current campaign, should go in the wordpress
as the page.