new friends page: Removed dynamic content and refactored static bits
This is major change which removed the part about the current campaign.
It is supposed to live in the wordpress database as that description is changing.

The things that are more code and do not change as much are the PayPal
bits. These are now uglily refactored. It's really messy. But it seems
to work. But right now it just echo()s stuff right onto the page. It'd
be better if the functions generated strings. But I don't know how to make
PHP do that easily.

It's been tested in a Wordpress virtual machine from BitNami. So it renders
correctly as far as I can see. But whether Paypal still works hasn't been tested
yet. Fortunately, we can now change the PayPal Action URL in one single place which
allows easier testing.

And if that should go live, it needs to be renamed to "page-friends.php".