FoG: A new 2012 donor
The problem was (or is) that the paypal export is weirdly encoded. It
presumably uses windows-1252 but iconv --from-code windows-1252
--to-code utf-8 doesn't decode the CSV file properly. Hence the name
didn't fall out when grepping through the export.

Right now we use information from the Instant Payment Notifications to
put names of donors on the list. This is hopefully more reliable.

Thanks to Jiro for making us aware of the missing name.